% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution hebdomadaire des trajets}
\usage{
calcul_distribution_semaine(trajet, filtre = TRUE)
}
\arguments{
\item{trajet}{Un data.frame contenant les données de trajets vélo.}

\item{filtre}{Booléen. Si TRUE, applique d'abord filtre_anomalie().}
}
\value{
Un data.frame avec deux colonnes :
\describe{
\item{jour_semaine}{Le numéro du jour de la semaine}
\item{nombre_trajets}{La somme des trajets pour ce jour}
}
}
\description{
Cette fonction calcule la somme des trajets par jour de la semaine.
Par défaut, elle filtre d'abord les anomalies.
}
