% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{compter_nombre_trajets}
\alias{compter_nombre_trajets}
\title{Compter le nombre de trajets}
\usage{
compter_nombre_trajets(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les données de trajets vélo.}
}
\value{
Un nombre entier correspondant au total de trajets.
}
\description{
Cette fonction calcule la somme de la colonne Total.
}
