% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_velo}
\alias{df_velo}
\title{Données de trajets vélo pendant les vacances de Toussaint 2025}
\format{
Un data.frame avec plusieurs colonnes décrivant les trajets :
\describe{
\item{date}{Date du trajet}
\item{heure}{Heure du trajet}
\item{boucle}{Numéro de boucle}
\item{nb_velos}{Nombre de vélos comptés}
}
}
\source{
Source du fichier téléchargé dans l'exercice
}
\usage{
df_velo
}
\description{
Ce jeu de données contient les trajets vélo observés à Nantes pendant
les vacances de Toussaint 2025.
}
\keyword{datasets}
