% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtre_anomalie}
\alias{filtre_anomalie}
\title{Filtrer les lignes considérées comme anormales}
\usage{
filtre_anomalie(trajet, seuil = 0.8)
}
\arguments{
\item{trajet}{Un data.frame contenant les données de trajets vélo.}

\item{seuil}{Un nombre entre 0 et 1. Les lignes avec une probabilité
d'anomalie strictement supérieure à ce seuil sont supprimées.}
}
\value{
Un data.frame filtré.
}
\description{
Cette fonction retire les lignes dont la probabilité de présence
d'anomalies dépasse un seuil donné.
}
