% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{filtrer_trajet}
\alias{filtrer_trajet}
\title{Filtrer les trajets selon un ou plusieurs numéros de boucle}
\usage{
filtrer_trajet(trajet, boucle)
}
\arguments{
\item{trajet}{Un data.frame contenant les données de trajets vélo.}

\item{boucle}{Un vecteur des numéros de boucle à conserver.}
}
\value{
Un data.frame filtré.
}
\description{
Cette fonction conserve uniquement les lignes dont le numéro de boucle
appartient au vecteur fourni.
}
