% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{plot_distribution_semaine}
\alias{plot_distribution_semaine}
\title{Tracer la distribution hebdomadaire des trajets}
\usage{
plot_distribution_semaine(trajet, filtre = TRUE)
}
\arguments{
\item{trajet}{Un data.frame contenant les données de trajets vélo.}

\item{filtre}{Booléen. Si TRUE, applique d'abord filtre_anomalie().}
}
\value{
Un graphique ggplot2.
}
\description{
Cette fonction produit un graphique du nombre total de trajets
selon le jour de la semaine.
}
