% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velo.R
\name{trouver_trajet_max}
\alias{trouver_trajet_max}
\title{Trouver la ligne correspondant au trajet maximal}
\usage{
trouver_trajet_max(trajet)
}
\arguments{
\item{trajet}{Un data.frame contenant les données de trajets vélo.}
}
\value{
Un data.frame d'une ligne correspondant au maximum de la colonne Total.
}
\description{
Cette fonction renvoie la ligne du data.frame ayant le Total maximal.
}
